Renamed list controller to trends controller
