+++ config/services/controller.yml
@@ -6,9 +6,9 @@ services:
             - '%allowed.origin%'
             - '%kernel.environment%'
 
-    App\Trends\Infrastructure\Controller\ListController:
+    App\Trends\Infrastructure\Controller\TrendsController:
         public: true
-        class: 'App\Trends\Infrastructure\Controller\ListController'
+        class: 'App\Trends\Infrastructure\Controller\TrendsController'
         arguments:
             - '%allowed.origin%'
             - '%kernel.environment%'
